+++ src/main/java/kakaotech/bootcamp/respec/specranking/chatconsumer/domain/chat/config/WebClientConfig.java
@@ -1,4 +1,4 @@
-package kakaotech.bootcamp.respec.specranking.chatconsumer.global.config;
+package kakaotech.bootcamp.respec.specranking.chatconsumer.domain.chat.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
