+++ src/main/java/kakaotech/bootcamp/respec/specranking/chatconsumer/domain/chat/dto/consume/ChatConsumeDto.java
@@ -1,6 +1,6 @@
 package kakaotech.bootcamp.respec.specranking.chatconsumer.domain.chat.dto.consume;
 
-import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.common.type.ChatStatus;
+import kakaotech.bootcamp.respec.specranking.chatconsumer.global.common.type.ChatStatus;
 
 public record ChatConsumeDto(
         String idempotentKey,
