+++ src/main/java/kakaotech/bootcamp/respec/specranking/chatconsumer/domain/chat/entity/Chat.java
@@ -11,8 +11,8 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.chatroom.entity.Chatroom;
-import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.common.BaseTimeEntity;
 import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.user.entity.User;
+import kakaotech.bootcamp.respec.specranking.chatconsumer.global.common.Entity.BaseTimeEntity;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
