+++ src/main/java/kakaotech/bootcamp/respec/specranking/chatconsumer/domain/chatparticipation/entity/ChatParticipation.java
@@ -9,8 +9,8 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.chatroom.entity.Chatroom;
-import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.common.BaseTimeEntity;
 import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.user.entity.User;
+import kakaotech.bootcamp.respec.specranking.chatconsumer.global.common.Entity.BaseTimeEntity;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
