--- src/main/java/kakaotech/bootcamp/respec/specranking/chatconsumer/domain/common/type/ChatStatus.java
@@ -1,5 +1,0 @@
-package kakaotech.bootcamp.respec.specranking.chatconsumer.domain.common.type;
-
-public enum ChatStatus {
-    SENT
-}
