--- src/main/java/kakaotech/bootcamp/respec/specranking/chatconsumer/domain/common/type/NotificationTargetType.java
@@ -1,6 +1,0 @@
-package kakaotech.bootcamp.respec.specranking.chatconsumer.domain.common.type;
-
-public enum NotificationTargetType {
-    CHAT,
-    SOCIAL
-}
