--- src/main/java/kakaotech/bootcamp/respec/specranking/chatconsumer/domain/common/type/UserRole.java
@@ -1,6 +1,0 @@
-package kakaotech.bootcamp.respec.specranking.chatconsumer.domain.common.type;
-
-public enum UserRole {
-    ROLE_USER
-}
-
