--- src/main/java/kakaotech/bootcamp/respec/specranking/chatconsumer/domain/common/type/UserStatus.java
@@ -1,7 +1,0 @@
-package kakaotech.bootcamp.respec.specranking.chatconsumer.domain.common.type;
-
-public enum UserStatus {
-    ACTIVE,
-    INACTIVE,
-    DELETED
-}
