+++ src/main/java/kakaotech/bootcamp/respec/specranking/chatconsumer/domain/notification/entity/Notification.java
@@ -10,8 +10,8 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.common.type.NotificationTargetType;
 import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.user.entity.User;
+import kakaotech.bootcamp.respec.specranking.chatconsumer.global.common.type.NotificationTargetType;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
