+++ src/main/java/kakaotech/bootcamp/respec/specranking/chatconsumer/domain/notification/repository/NotificationRepository.java
@@ -1,7 +1,7 @@
 package kakaotech.bootcamp.respec.specranking.chatconsumer.domain.notification.repository;
 
-import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.common.type.NotificationTargetType;
 import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.notification.entity.Notification;
+import kakaotech.bootcamp.respec.specranking.chatconsumer.global.common.type.NotificationTargetType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface NotificationRepository extends JpaRepository<Notification, Long> {
