+++ src/main/java/kakaotech/bootcamp/respec/specranking/chatconsumer/domain/user/entity/User.java
@@ -9,9 +9,9 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Index;
 import jakarta.persistence.Table;
-import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.common.BaseTimeEntity;
-import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.common.type.UserRole;
-import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.common.type.UserStatus;
+import kakaotech.bootcamp.respec.specranking.chatconsumer.global.common.Entity.BaseTimeEntity;
+import kakaotech.bootcamp.respec.specranking.chatconsumer.global.common.type.UserRole;
+import kakaotech.bootcamp.respec.specranking.chatconsumer.global.common.type.UserStatus;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
