+++ src/main/java/kakaotech/bootcamp/respec/specranking/chatconsumer/global/common/type/ChatStatus.java
@@ -1,0 +1,5 @@
+package kakaotech.bootcamp.respec.specranking.chatconsumer.global.common.type;
+
+public enum ChatStatus {
+    SENT
+}
