+++ src/main/java/kakaotech/bootcamp/respec/specranking/chatconsumer/global/common/type/NotificationTargetType.java
@@ -1,0 +1,6 @@
+package kakaotech.bootcamp.respec.specranking.chatconsumer.global.common.type;
+
+public enum NotificationTargetType {
+    CHAT,
+    SOCIAL
+}
