+++ src/main/java/kakaotech/bootcamp/respec/specranking/chatconsumer/global/common/type/UserStatus.java
@@ -1,0 +1,7 @@
+package kakaotech.bootcamp.respec.specranking.chatconsumer.global.common.type;
+
+public enum UserStatus {
+    ACTIVE,
+    INACTIVE,
+    DELETED
+}
